python_scripts/requirements.txt
openai==1.3.0
python-dotenv==1.0.0 